test: comafi_accounts: fix to pass the integration test
The integration test is expecting to return the path of the result
of the comafi account preparation process, but it was returnining a string
with part of the name of the directory.
Also delete a print that was while developing and nnecessary.
Rename the name of new_folder to new_directory. more linux :)